// src/server/model/iLedger.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IAttributeGroupLine } from './iAttributeGroupLine';
import { IContact } from './iContact';
import { IDocument } from './iDocument';
import { ILedgerLabel } from './iLedgerLabel';
import { IOpeningBalanceTx } from './iOpeningBalanceTx';


export interface ILedger { 
    disabled?: boolean;
    labels?: Array<ILedgerLabel>;
    printname?: string;
    gstin?: string;
    cstNumber?: string;
    panNumber?: string;
    aliasName?: string;
    gstStatus?: ILedger.GstStatusEnum;
    unit?: string;
    costCenter?: string;
    contact?: IContact;
    salesManId?: number;
    gstType?: ILedger.GstTypeEnum;
    tinNumber?: string;
    profitShare?: number;
    contactAddress?: string;
    ledgerGroupId?: number;
    ledgerGroupName?: string;
    serviceTaxNumber?: string;
    openingBalance?: number;
    attrGroupLine?: IAttributeGroupLine;
    depreciationRate?: number;
    noOfCreditDays?: number;
    hasCostCenter?: boolean;
    salesManName?: string;
    openingBalTx?: IOpeningBalanceTx;
    verificationDate?: Date;
    documentList?: Array<IDocument>;
    amountCreditLimits?: number;
    location?: string;
    type?: number;
    systemGenerated?: boolean;
    name?: string;
    description?: string;
    importedId?: number;
    logInfo?: string;
    productUserId?: number;
    creationdate?: Date;
    productUserName?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    id?: number;
    jacksontype?: string;
}
export namespace ILedger {
    export type GstStatusEnum = 'VALID' | 'INVALID' | 'NOT_VERIFIED';
    export const GstStatusEnum = {
        VALID: 'VALID' as GstStatusEnum,
        INVALID: 'INVALID' as GstStatusEnum,
        NOTVERIFIED: 'NOT_VERIFIED' as GstStatusEnum
    };
    export type GstTypeEnum = 'Regular' | 'Composition' | 'Unregistered';
    export const GstTypeEnum = {
        Regular: 'Regular' as GstTypeEnum,
        Composition: 'Composition' as GstTypeEnum,
        Unregistered: 'Unregistered' as GstTypeEnum
    };
}